Assume all sessions removed on PrepareForShutdown.
SessionRemoved is not sent in time if the user shuts down without
explicitly logging out and waiting ~15 seconds first. As a stop-gap,
make the aggressive assumption that all sessions are removed when the
PrepareForShutdown signal is sent. This assumption does not hold when
the shutdown is cancelled, but presumably that is a much less common
event than the user shutting down without explicitly logging out.

[endlessm/eos-shell#2216]