Added userID to login metric.
The login (and thus logout) metric now tracks the user id via the
auxiliary payload when the login is registered.

The instrumentation code is itself responsible for the defintion
of the login/logout metric.  This pattern may or may not be extended
to all non-metrics repos in the near future.
[endlessm/eos-sdk#1686]