Record StartupFinished event.
Record the StartupFinished signal emitted by the systemd manager DBus
interface. Ferry the parameters of the signal through to the event
recorder, remaining agnostic to their contents. These parameters provide
a breakdown of the amount of time spent in each phase of the boot. More
details are available at
http://www.freedesktop.org/wiki/Software/systemd/dbus/.

Note that the actual time at which this event is recorded is not
semantically significant even though that data is sent to the server.

[endlessm/eos-sdk#638]