Record user ID of user that logged in.
Currently we populate the auxiliary payload of the login event with the
user ID of the metrics instrumentation daemon. The metrics
instrumentation daemon always runs as root in production, so the user
ID we are recording is always 0. This change instead populates the
auxiliary payload of the login event with the user ID of the user that
logged in.

[endlessm/eos-sdk#1686]