Launch after event recorder daemon.
Since the metrics instrumentation daemon requires the event recorder
daemon in order to record any events, we make it list the event
recorder daemon as a dependency and ensure that the metrics
instrumentation daemon launches after and closes before the event
recorder daemon. This also helps the shutdown event have enough time to
be persisted to disk by the event recorder daemon since it is recorded
when the instrumentation daemon receives the SIGTERM signal from
systemd.

[endlessm/eos-sdk#2906]