Allow overriding default location of tally file
In order to run integration tests on the daemon, it must not try to write
to a file in $(localstatedir). Normally when $(localstatedir) is /var,
this would produce a harmless warning. If $(localstatedir) is writable,
it leaves a useless file lying around. However, during make distcheck,
the file is written into the install tree, but is not uninstalled by make
uninstall, and so make distcheck fails.

[endlessm/eos-sdk#861]