Avoid unnecessary error handling.
In general we don't want to attempt error handling in cases we never
expect to enter because they are very difficult to test. If the
persistent tally key isn't found even though the key file exists, then
most likely either there is a programmer error that we shouldn't mask,
or the key file has been tampered with, and we have little hope of
redemption.

[endlessm/eos-sdk#3040]