Fix handling of multiple keys in persistent tally.
Previously persistent tally objects were 1:1 with keys in key files, but
this led to race conditions when multiple objects were backed by the
same key file. As a consequence, the boot count incremented very rarely.
Since the boot count is unusable before this commit, we assign a new
event ID to the uptime and shutdown events. This will allow us to
distinguish between events with invalid boot counts and events with
valid boot counts on the server.

[endlessm/eos-sdk#3848]