retrieve eos-image-version xattr from /sysroot and send as a metric
This adds event 6b1c1cfc-bc36-438c-0647-dacd5878f2b3 which is sent once
at startup to report the image ID. This is a string such as
"eos-eos3.1-amd64-amd64.170115-071322.base" which is saved in an
attribute on the root filesystem by the image builder, and allows us
to tell the channel that the OS was installed by (eg download, OEM
pre-install, Endless hardware, USB stick, etc) and which version was
installed. The payload is a single string containing this image ID,
if present.