data: delay gathering metrics until FBE is complete
The metrics daemon is disabled by default. The first-boot experience
shows a checkbox asking the user whether to enable metrics, which
defaults to checked. At the time the FBE loads, the checkbox state is
synced to the metrics daemon, enabling it.

eos-metrics-instrumentation currently runs before the FBE loads. As a
result, the first set of boot-time metrics are discarded. On live boots,
this is equivalent to all boot-time metrics being discarded.

Instead, we use a path unit to trigger boot-time metrics collection when
/home is non-empty. On the first boot, this will be true when the FBE is
completed; on subsequent boots, this will be true at boot time.

This is the same technique we use for eos-phone-home.

https://phabricator.endlessm.com/T14065