implement network ID and eos-network-id utility
The network ID is stable opaque identifier which should be the same
for machines which are connected to the same physical network. It is
an unsigned 32-bit hash of the ethernet MAC address of the gateway
(aka router) that provides the default route on IPv4 (if available)
or IPv6 (if not).

The eos-network-id utility determines this ID for the current system
and prints it.