record event with network ID
Introduce the event "38eb48f8-e131-9b57-77c6-35e0590c82fd" which is
recorded at system startup and when we detect a change in the default
route. The auxiliary payload is a 32-bit unsigned integer containing
a hash of the ethernet MAC address of the gateway, favouring IPv4 if
available, or IPv6 if not. The intention is to provide a value which
is opaque and stable which is the same for every system located on
the same physical network.