record event with location label
Introduce the event "eb0302d8-62e7-274b-365f-cd4e59103983" which is recorded
at system startup and whenever location.conf is modified (usually by the
system administrator using the eos-label-location command). The auxiliary
payload is a dictionary of string keys (such as facility, city and state)
to the values provided in the location.conf file. The intention is to allow
an operator to provide an optional human-readable label for the location of
the system, which can be used when preparing reports or visualisations of the
metrics data.