eos-crash-metrics: better information for Flatpak applications
Looking at the executable path we can make an assunmption whether it is
a flatpak'ed application that crashed or not. Then we compare
that executable name with the ones from the installed flatpaks
on the system.

We use the flatpak library API then to get information on the
full ref, the commit and the repo url of the app and the
corresponding runtime.

The flatpak information is optional, so we can still do the
metrics for non flatpak'ed apps.

https://phabricator.endlessm.com/T19194