Implement CPU info in terms of lscpu
On ARM, the 'model name' field of /proc/cpuinfo contains only very vague
model information. More useful information is encoded in the 'CPU
implementer' and 'CPU part' fields. For example, here's an excerpt from
/proc/cpuinfo on an ARM server:

        processor	: 0
        model name	: ARMv7 Processor rev 0 (v7l)
        BogoMIPS	: 2791.42
        Features	: half thumb fastmult vfp edsp thumbee neon vfpv3 tls vfpd32
        CPU implementer	: 0x41
        CPU architecture: 7
        CPU variant	: 0x3
        CPU part	: 0xc09
        CPU revision	: 0

glibtop_get_sysinfo() on Linux just parses each paragraph of
/proc/cpuinfo into a hash table, so unless we want to decode the
implementer/part fields ourselves (whether on the client, or on the
server) then we need to look for an alternative solution.

Recent versions of lscpu from util-linux contain a [lookup table][]
which decodes the implementer/part fields into the human-readable
product name. For example, it can determine that the CPU above is in
fact an ARM Cortex-A9.

Let's reuse this lookup table by parsing the output from lscpu rather
than using libgtop. (There's an argument that libgtop should be made to
do the same so that a more-informative human-readable string is
displayed in gnome-control-center.)

[lookup table]: https://github.com/karelzak/util-linux/commit/744d62ee

https://phabricator.endlessm.com/T18445