Only record city on Solutions systems
The user's city can be useful for many benign analysis purposes, but
since it can be correlated with all of the other metrics via the machine
ID, it transforms a lot of other data into personally-identifying
information that otherwise wouldn't necessarily have to be considered as
such.

We stop recording it in general. However, it may still be needed to keep
track of Solutions deployments, so on Solutions images we continue
recording it.

Since the main eos-metrics-instrumentation daemon now doesn't record the
location unconditionally, we can't use it in the test script. Instead,
the test script starts a special binary that always records the location
so that the test bus can verify the input and output of these functions.

https://phabricator.endlessm.com/T27349