Stop sending NETWORK_STATUS_CHANGED_EVENT
This event (5fae6179-e108-4962-83be-c909259c0584) is the most
frequently-submitted event, with 500 million submissions in a year.
It's not particularly valuable, given that we have never had cause nor
desire to try and analyse the data in the 4 years since I joined
Endless. Stop sending it to save space and bandwidth, as the new metrics
system will also ignore and not store it.

https://phabricator.endlessm.com/T28301