Fix double free in /proc/cmdline error path
The 'error' variable is declared as g_autoptr(GError), meaning that if
it is non-NULL when it goes out of scope, g_error_free() will be called
on it. On the error path, we currently explicitly call g_error_free(),
and do not clear 'error' after doing so. This is a double-free.

Fortunately, reading /proc/cmdline never fails. Nonetheless, delete the
offending line.