location-label: Ignore empty values
If you use eos-label-location and don't give an answer to one or more of
its questions, you get a file like this:

    [Label]
    facility = Will's House
    city = London
    state =

The 'site' page in our branch of gnome-initial-setup similarly writes
empty files like this:

    [Label]
    id =
    city =
    state =
    street =
    country =
    facility =

As a result of this second case, over 99% of the rows in our production
table for this event are:

    {"id": "", "city": "", "state": "", "street": "", "country": "", "facility": ""}

Handle the first case by ignoring empty fields. Handle the second case
by not reporting a metric if there are no non-empty fields. Test these
cases.

There is an edge case that is not covered by this change: if a machine
previously has non-empty information in this file, and is updated to
have only empty keys in the file (or no file at all), no event is
submitted. This will cause the machine to be incorrectly assigned to its
old location information on the server.

This is believed to be rare.

https://phabricator.endlessm.com/T30469