Assert when stopping a session timer
g_hash_table_remove() returns TRUE if something was removed and FALSE
otherwise. We always expect the user that has logged out to have had a
timer in this hash table, so we can assert that it returns TRUE.

(Might this assertion fail in the case where eos-metrics-instrumentation
starts up after a user logs in, but that user logs out before it has
finished enumerating the logged-in users? I believe the answer is no,
because it enumerates synchronously and this callback will not fire
before returning to the mainloop.)

https://phabricator.endlessm.com/T33295