Don't assert when inhuman users log out
The assertion added in 8c7a7e74c5d36a1ef25f0a7aaf2448ba0a642d59 ("Assert
when stopping a session timer") was invalid.

add_session() only adds a session to the session_by_user_id hash table
if its user_id is in the range assigned to humans. But remove_session()
would assert that a now-ended session was in the hash table regardless
of its user_id. One example of such an inhuman user is Debian-gdm, uid
109, which logs out when the greeter is killed to release resources
while a user is logged in. As a result, eos-metrics-instrumentation
would reliably abort shortly after the first user logs in.

Add the same guard as found in add_session() to remove_session().

https://phabricator.endlessm.com/T33295