Soften session timer assertion to a warning
I believe that, since the fix in the previous commit, the assertion will
now never fail unless logind is buggy.

However, I also believed that before the previous commit; and it is
possible that logind is buggy, or the signals are not emitted in pairs
as this code assumes based on its documentation and observed behaviour.

Replace the assertion with a warning, providing more information as to
what unexpected condition has occurred.

https://phabricator.endlessm.com/T33295