hwinfo: Fix path to record_time file
In the Autotools build system (removed in
2751d4a329df06f573cacf455b5fc5801a7cec9a) there was a trailing slash on
INSTRUMENTATION_CACHE_DIR. In the Meson build system (added in
96dab569cfe4c11b6f6a5ae59713c396bfd9e62e) there is not. As a result,
this daemon has been writing to
/var/cache/eos-metrics-instrumentationrecord_time rather than
/var/cache/eos-metrics-instrumentation/record_time.

Fix this by adding a separator when the strings are concatenated.

Because this change has not entered a stable release of Endless OS, I'm
not going to worry about deleting or moving the erroneous file.

https://phabricator.endlessm.com/T33592