hwinfo: Force C numeric locale for lscpu
I noticed that the data submitted every day by my laptop has wildly
varying values for the max frequency field:

    azafea=> select occured_at :: date as "Day",  info -> 0 ->> 'model' as "Model", info -> 0 ->> 'max_frequency' as "Max Frequency" from computer_information_v3 where channel_id = 1 order by occured_at desc limit 10;
        Day     |                  Model                   | Max Frequency
    ------------+------------------------------------------+---------------
     2022-07-03 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 1513.099
     2022-07-02 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 3000.002
     2022-07-01 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 886.474
     2022-06-30 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2000.0
     2022-06-29 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2000.0
     2022-06-28 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2400.05
     2022-06-27 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2299.999
     2022-06-24 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2000.0
     2022-06-23 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2000.0
     2022-06-22 | Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz | 2000.0
    (10 rows)

(I have tagged my computer with a unique site ID so I can uniquely
identify it in the otherwise-anonymous metrics data.)

This is because my system locale is French, and the output of 'lscpu
--json' is locale-sensitive. In a locale like French, the min and max
CPU fields use ',' not '.' as the decimal separator.

This wasn't caught by test_get_computer_hwinfo(), which checks that
eins_hwinfo_get_computer_hwinfo() produces somewhat reasonable output on
the current system, because lscpu is inconsistent!

{
   "lscpu": [
      {"field":"Model name:", "data":"Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz"},
      {"field":"CPU MHz:", "data":"2000.000"},
      {"field":"CPU max MHz:", "data":"4600,0000"},
      {"field":"CPU min MHz:", "data":"400,0000"},
      {"field":"BogoMIPS:", "data":"3999.93"}
   ]
}

So this code would fail to parse "CPU max MHz:", log a debug message,
and fall back to "CPU MHz:", which it successfully parsed. So it would
not use the ultimate fallback, 0, which would cause the test to fail.
And there is no non-zero value that we can compare against.

This is fixed in more recent util-linux than we have in Endless OS
today. https://github.com/util-linux/util-linux/issues/1743

I tried to construct a failing test case by raising the severity of the
debug message to a warning, then forcing LC_NUMERIC to French for this
test, but the problem is that there are other tests which trigger the
"parse failed" path, so the warning made those fail. (We could adjust
those tests but I feel like we are in diminishing returns territory
here.)

`man lscpu` does have the following to say:

    MAXMHZ

    Maximum megahertz value for the CPU. Useful when lscpu is used as
    hardware inventory information gathering tool.  Notice that the
    megahertz value is dynamic, and driven by CPU governor depending on
    current resource need.

So, this value is more dynamic than I expected. Still, it is less
dynamic than the current CPU frequency, which is highly variable. The
nominal frequency from the Model name: field is unfortunately not
exposed in any easily-accessed field in the output.

https://phabricator.endlessm.com/T33657