Run as metrics user, not root
This applies to the main daemon, not the coredump handler.

I have been running with this change locally for some time and have
noticed no adverse effects. An unprivileged user can receive systemd's
signals about startup finishind and users logging in and out, and
execute `lscpu --json` with no difference in output.

Perhaps in the past there was a reason to run as root, but no longer.

https://phabricator.endlessm.com/T9267