debian: Build-depend on systemd-dev
systemd.pc, which is used to look up the paths to install a systemd unit,
tmpfiles.d snippet and sysctl.d snippet, is part of this package.

When we updated systemd in EOS 6, we dropped our patch for the the `systemd`
binary package in previous branches which marked it as `Priority: required`,
instead leaving it as `Priority: important`. When it was `required`, it would
always be present in the buildroot, and so would `systemd-dev` which it
(surprisingly!) depends on; but now it is not. 

Explicitly declare this build dependency, as we should have been doing all along.

https://phabricator.endlessm.com/T35070#990194
https://phabricator.endlessm.com/T35352